Fix Go code formating
