+++ sonar-go-checks/src/test/resources/checks/TooDeeplyNestedStatementsCheck/too_deeply_nested_statements_max_2.go
@@ -1,17 +1,17 @@
 package TooDeeplyNestedStatementsCheck
 
 func main() {
-	if true {
-		//^^>
-	} else if true {
-	} else if true {
-		if true { // Compliant
-			//  ^^>
-			if true { // Noncompliant {{Refactor this code to not nest more than 2 control flow statements.}}
-				//    ^^
-				if true {
-				}
-			}
-		}
-	}
+  if true {
+//^^>
+  } else if true {
+  } else if true {
+    if true { // Compliant
+//  ^^>
+      if true { // Noncompliant {{Refactor this code to not nest more than 2 control flow statements.}}
+//    ^^
+        if true {
+        }
+      }
+    }
+  }
 }
